# server/swagger.json
{
  "swagger": "2.0",
  "info": {
    "description": "This is the Swagger Documentation of SafeWorks server API",
    "version": "1.0.0",
    "title": "SafeWorks API"
  },
  "host": "localhost:3001",
  "basePath": "/api",
...
  "schemes": [
    "http",
    "https"
  ],
  "consumes": [
    "application/json"
  ],
  "produces": [
    "application/json"
  ],
  "paths": {
...
    "/users": {
      "post": {
        "tags": [
          "User"
        ],
        "description": "Endpoint to add User",
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "name": "user",
            "in": "body",
            "description": "User data to add",
            "required": true,
            "schema": {
              "$ref": "#/definitions/AddUser"
            }
          }
        ],
        "responses": {
          "201": {
            "description": "Created user JWT token",
            "schema": {
              "$ref": "#/definitions/CreatedUser"
            }
          },
          "409": {
            "description": "Email already exists",
            "schema": {
              "$ref": "#/definitions/ConflictEmail"
            }
          },
          "500": {
            "description": "Action fail",
            "schema": {
              "$ref": "#/definitions/FailedAction"
            }
          }
        }
      }
    },
    "/users/{uid}": {
      "get": {
        "tags": [
          "User"
        ],
        "description": "Endpoint to get a User by ID",
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "name": "uid",
            "in": "path",
            "required": true,
            "type": "string",
            "description": "User UID"
          }
        ],
        "responses": {
          "200": {
            "description": "User response object",
            "schema": {
              "$ref": "#/definitions/User"
            }
          },
          "404": {
            "description": "Not Found"
          },
          "500": {
            "description": "Action fail",
            "schema": {
              "$ref": "#/definitions/FailedAction"
            }
          }
        }
      },
      "put": {
        "tags": [
          "User"
        ],
        "description": "Endpoint to update a User",
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "name": "uid",
            "in": "path",
            "required": true,
            "type": "string",
            "description": "User UID"
          },
          {
            "name": "user",
            "in": "body",
            "description": "User data to update",
            "required": true,
            "schema": {
              "$ref": "#/definitions/AddUser"
            }
          }
        ],
        "responses": {
          "201": {
            "description": "Updated user JWT token",
            "schema": {
              "$ref": "#/definitions/CreatedUser"
            }
          },
          "404": {
            "description": "Not Found"
          },
          "500": {
            "description": "Action fail",
            "schema": {
              "$ref": "#/definitions/FailedAction"
            }
          }
        }
      }
    },
    "/users/login": {
      "post": {
        "tags": [
          "User"
        ],
        "description": "Endpoint to login a User",
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "name": "user",
            "in": "body",
            "description": "User data to login",
            "required": true,
            "schema": {
              "$ref": "#/definitions/LoginUser"
            }
          },
          {
            "name": "accessToken",
            "in": "header",
            "description": "User access token to login",
            "required": false,
            "type": "string"
          }
        ],
        "responses": {
          "400": {
            "description": "User login temporarily disabled"
          },
          "401": {
            "description": "Logged user JWT token",
            "schema": {
              "$ref": "#/definitions/LoggedUser"
            }
          },
          "500": {
            "description": "Action fail",
            "schema": {
              "$ref": "#/definitions/FailedAction"
            }
          }
        }
      }
    },
    "/users/upload": {
      "post": {
        "tags": [
          "User"
        ],
        "description": "Endpoint to upload a User image file",
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "name": "file",
            "in": "formData",
            "type": "file",
            "description": "Image file to upload",
            "required": true
          },
          {
            "name": "uid",
            "in": "formData",
            "type": "string",
            "description": "User UID",
            "required": true
          }
        ],
        "responses": {
          "200": {
            "description": "Uploaded user photo url",
            "schema": {
              "$ref": "#/definitions/UploadedUserPhoto"
            }
          },
          "404": {
            "description": "Not Found"
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      }
    }
  },
  "definitions": {
    "User": {
      "type": "object",
      "properties": {
        "statusCode": {
          "type": "number",
          "example": 200
        },
        "user": {
          "type": "object",
          "properties": {
            "uid": {
              "type": "string",
              "example": "SXn2duju1bS6uXRZBKgXoVbC43D2"
            },
            "email": {
              "type": "string",
              "example": "lucas@email.com"
            },
            "password": {
              "type": "string",
              "example": "@Password1234"
            },
            "name": {
              "type": "string",
              "example": "Lucas"
            },
            "username": {
              "type": "string",
              "example": "lucaskleal"
            },
            "cpf": {
              "type": "string",
              "example": "45644283062"
            },
            "telephone_number": {
              "type": "string",
              "example": "41935610501"
            },
            "district": {
              "type": "string",
              "example": "Centro"
            },
            "photo_url": {
              "type": "string",
              "example": "https://1231245"
            },
            "balance": {
              "type": "number",
              "example": 0
            },
            "contracted_services": {
              "type": "array",
              "example": [],
              "items": {}
            },
            "worker": {
              "type": "object",
              "properties": {
                "area": {
                  "type": "array",
                  "example": [
                    "Carpinteiro"
                  ],
                  "items": {
                    "type": "string"
                  }
                },
                "portfolio": {
                  "type": "string",
                  "example": "huHwBD3gTccU7pgA3Z94"
                }
              }
            },
            "created": {
              "type": "string",
              "example": "19/06/2023, 18:23:02"
            },
            "modified": {},
            "deleted": {},
            "accessToken": {
              "type": "string",
              "example": "fgjdfgldfgprk"
            }
          }
        }
      }
    },
    "AddUser": {
      "type": "object",
      "properties": {
        "email": {
          "type": "string",
          "example": "lucas@email.com"
        },
        "password": {
          "type": "string",
          "example": "@Password1234"
        },
        "name": {
          "type": "string",
          "example": "Lucas"
        },
        "cpf": {
          "type": "string",
          "example": "45644283062"
        },
        "telephone_number": {
          "type": "string",
          "example": "41935610501"
        }
      },
      "required": [
        "email",
        "password",
        "name",
        "cpf",
        "telephone_number"
      ]
    },
    "CreatedUser": {
      "type": "object",
      "properties": {
        "statusCode": {
          "type": "number",
          "example": 201
        },
        "token": {
          "type": "string",
          "example": "asda231234afasfg3dsfsd"
        }
      }
    },
    "ConflictEmail": {
      "type": "object",
      "properties": {
        "statusCode": {
          "type": "number",
          "example": 409
        },
        "error": {
          "type": "string",
          "example": "email/already-exists"
        },
        "message": {
          "type": "string",
          "example": "Email already exits..."
        }
      }
    },
    "LoginUser": {
      "type": "object",
      "properties": {
        "email": {
          "type": "string",
          "example": "lucas@email.com"
        },
        "password": {
          "type": "string",
          "example": "@Password1234"
        }
      },
      "required": [
        "email",
        "password"
      ]
    },
    "LoggedUser": {
      "type": "object",
      "properties": {
        "statusCode": {
          "type": "number",
          "example": 401
        },
        "token": {
          "type": "string",
          "example": "asda231234afasfg3dsfsd"
        }
      }
    },
    "UploadedUserPhoto": {
      "type": "object",
      "properties": {
        "statusCode": {
          "type": "number",
          "example": 200
        },
        "image": {
          "type": "string",
          "example": "https://SXn2duju1bS6uXRZBKgXoVbC43D2_profile_photo"
        }
      }
    },
...
    "FailedAction": {
      "type": "object",
      "properties": {
        "statusCode": {
          "type": "number",
          "example": 500
        },
        "error": {
          "type": "string",
          "example": "model/failed-action"
        },
        "message": {
          "type": "string",
          "example": "Error on action..."
        }
      }
    }
  }
}